Move locale plug to separate file
